--- a/DevOpsMetrics/DevOpsMetrics.Tests/Service/ChangeFailureRateDATests.cs
+++ b/DevOpsMetrics/DevOpsMetrics.Tests/Service/ChangeFailureRateDATests.cs
@@ -3,6 +3,7 @@ using DevOpsMetrics.Service.Models.Common;
 using Microsoft.Extensions.Configuration;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace DevOpsMetrics.Tests.Service