--- a/DevOpsMetrics/DevOpsMetrics.Web/Views/Home/ChangeFailureRateUpdates.cshtml
+++ b/DevOpsMetrics/DevOpsMetrics.Web/Views/Home/ChangeFailureRateUpdates.cshtml
@@ -28,7 +28,7 @@
         <div class="col-md-6">
             <span>Number of days to process:</span>
             <select asp-for="NumberOfDaysSelected" asp-items="Model.NumberOfDaysList" class="form-control">
-                <option value="">Select completion %</option>
+                <option value="">Select days to process</option>
             </select>
         </div>
     </div>