Improving performance for a few metrics that maybe could run faster

**Before**: (Based on DevOpsMetrics repo, over 30 days)
| Metric | before timing | after timing |
| -- | -- | -- |
| Deployment Frequency | 2.59s | 1.30s |
| Lead time for changes | 7.78s |
| Time to restore service | 0.67s | 0.25s |
| Change failure rate | 0.76s | 0.31s |